requirements.txt
cffi==1.15.0
ffmpeg-python==0.2.0
future==0.18.2
pycparser==2.21
vosk==0.3.32
